+++ R/param_em.R
@@ -91,8 +91,7 @@
 #'                      n_sub = 3),
 #'   WEIGHT = split_comp(add_var(NULL, data_1[, c("WEIGHT", "HEIGHT")]), n_sub = 2)
 #' )
-#' res_learn_1 <- param_em(gmbn_1, data_1, verbose = TRUE, regul = 0.01,
-#'                         max_iter_em = 100)
+#' res_learn_1 <- param_em(gmbn_1, data_1, verbose = TRUE)
 #'
 #' library(dplyr)
 #' set.seed(0)
@@ -129,14 +128,13 @@
 #'     WIND = split_comp(add_var(NULL, data_3[, c("WIND", "WIND.1")]), n_sub = 3)
 #'   )
 #' )
-#' res_learn_2 <- param_em(gmdbn_1, data_2, col_seq = "DATE", verbose = TRUE,
-#'                         regul = 0.01, max_iter_em = 100)}
+#' res_learn_2 <- param_em(gmdbn_1, data_2, col_seq = "DATE", verbose = TRUE)}
 #'
 #' @export
 
 param_em <- function(gmgm, data, nodes = structure(gmgm)$nodes, col_seq = NULL,
                      n_part = 1000, max_part_sim = 1e06, min_ess = 1,
-                     max_iter_pem = Inf, verbose = FALSE, ...) {
+                     max_iter_pem = 5, verbose = FALSE, ...) {
   is_gmbn <- gmgm %>%
     inherits("gmbn")