+++ tests/testthat/test-AIC.R
@@ -33,11 +33,11 @@ class(gmdbn_2) <- "gmdbn"
 
 test_that("compute the AIC of a gmm object", {
   expect_equal(AIC(gmm_1, data.frame(A = c(0, 3, 6, 9), B = c(4, 7, 1, 6))),
-               - 121.0873, tolerance = 0.01)
+               - 121.1373, tolerance = 0.01)
   expect_equal(AIC(gmm_1,
                    matrix(c(0, 3, 6, 9, 4, 7, 1, 6), 4,
                           dimnames = list(NULL, c("A", "B")))),
-               - 121.0873, tolerance = 0.01)
+               - 121.1373, tolerance = 0.01)
 })
 
 test_that("compute the AIC of a gmm object with missing values", {
@@ -49,18 +49,18 @@ test_that("compute the AIC of a gmm object with extra columns", {
   expect_equal(AIC(gmm_1,
                    data.frame(A = c(0, 3, 6, 9), B = c(4, 7, 1 , 6),
                               C = c(0, 0, 0, 0))),
-               - 121.0873, tolerance = 0.01)
+               - 121.1373, tolerance = 0.01)
 })
 
 test_that("compute the AIC of a gmm object with no row", {
-  expect_equal(AIC(gmm_1, data.frame(A = numeric(), B = numeric())), - 11)
-  expect_equal(AIC(gmm_1, data.frame(A = logical(), B = logical())), - 11)
+  expect_equal(AIC(gmm_1, data.frame(A = numeric(), B = numeric())), - 11.05)
+  expect_equal(AIC(gmm_1, data.frame(A = logical(), B = logical())), - 11.05)
 })
 
-test_that("compute the regularized AIC of a gmm object", {
+test_that("compute the non-regularized AIC of a gmm object", {
   expect_equal(AIC(gmm_1, data.frame(A = c(0, 3, 6, 9), B = c(4, 7, 1, 6)),
-                   regul = 0.01),
-               - 121.1373, tolerance = 0.01)
+                   regul = NULL),
+               - 121.0873, tolerance = 0.01)
 })
 
 test_that("compute the conditional AIC of a gmm object", {