+++ NEWS.md
@@ -1,3 +1,13 @@
+## gmgm (development version)
+
+* Changed the default values of several arguments to increase the speed of learning functions and make them easier to use:
+  * `regul = NULL` → `regul = 0.01` (functions `AIC`, `BIC`, `logLik`, `em`)
+  * `max_iter_em = Inf` → `max_iter_em = 100` (function `em`)
+  * `max_iter_smem = Inf` → `max_iter_smem = 10` (function `smem`)
+  * `max_iter_step = Inf` → `max_iter_step = 10` (function `stepwise`)
+  * `max_iter_pem = Inf` → `max_iter_pem = 5` (functions `param_em`, `struct_em`)
+  * `max_iter_sem = Inf` → `max_iter_sem = 5` (function `struct_em`)
+
 ## gmgm 1.0.2
 
 * Replaced dplyr's scoped verbs (`_all`, `_at`, `_if`), which have been superseded by the use of function `across`.